refactor: 기존 Redux 훅 제거 및 상태 모듈을 Zustand로 전환
